Add security group templates, restructure deploy for 3 phases

+++ solution/deploy.sh
@@ -11,16 +11,34 @@ _cleanup()
 
 trap _cleanup 0 1 2 3
 
-LIST=$(ls -1 *.yaml)
-COUNT=$(ls -1 *.yaml | wc -l)
-
-FIRST=$(ls -1 *.yaml | head -1)
-ALLBUTFIRST=$(ls -1 *.yaml | tail -n +2)
-
-STACK=cfn-demo
+# Defaults
+NETWORK=cfn-network
+SECGRP=cfn-secgrp
+PARENT=cfn-nested
 REGION=ap-southeast-1
+ACCOUNT=$(aws sts get-caller-identity --query Account --output text)
+BUCKET=${PARENT}-${REGION}-${ACCOUNT}
 
+if [[ -z "$ACCOUNT" ]]
+then
+    _error "Unable to determine AWS account"
+fi
 
+# Network (single stack) templates
+NETWORK_FIRST=$(ls -1 network-*.yaml | head -1)
+ALLBUT_NETWORK_FIRST=$(ls -1 network-*.yaml | tail -n +2)
+NETWORK_LAST=$(ls -1 network-*.yaml | tail -1)
+
+# SecGrp (additional non-nested stack) templates
+SECGRP_FIRST=$(ls -1 secgrp-*.yaml | head -1)
+ALLBUT_SECGRP_FIRST=$(ls -1 secgrp-*.yaml | tail -n +2)
+SECGRP_LAST=$(ls -1 secgrp-*.yaml | tail -1)
+
+# Parent (multi-stack) templates
+PARENT_FIRST=$(ls -1 parent-*.yaml | head -1)
+ALLBUTPARENT_FIRST=$(ls -1 parent-*.yaml | tail -n +2)
+
+# Colour output
 COLOR_RUN=$(tput setaf 12)
 COLOR_HEAD=$(tput setaf 10)
 COLOR_ERR=$(tput setaf 1)
@@ -78,7 +96,66 @@ function _error()
     exit 1
 }
 
-# https://github.com/alestic/aws-cloudformation-stack-status/blob/master/aws-cloudformation-stack-status
+# Usage: bucket options
+function _create_bucket()
+{
+    if [ $# -eq 0 ]
+    then
+        _error "_create_bucket Requires at least the bucket name"
+        exit 1
+    fi
+
+    B=$1
+    shift
+    EXTRAS="$*"
+
+    if aws s3api head-bucket --bucket $B $EXTRAS 2>&1 | grep -q 404;
+    then
+        _walk aws s3 mb s3://$B $EXTRAS > /dev/null
+        sleep 2
+    fi
+}
+
+# Usage: bucket options
+function _empty_bucket()
+{
+    if [ $# -eq 0 ]
+    then
+        _error "_empty_bucket Requires at least the bucket name"
+    fi
+
+    B=$1
+    shift
+    EXTRAS="$*"
+
+    if ! aws s3api head-bucket --bucket $B $EXTRAS 2>&1 | grep -q 404;
+    then
+        _run aws s3 rm --recursive s3://$B $EXTRAS > /dev/null
+    fi  
+}
+
+# Usage: bucket options
+function _delete_bucket()
+{
+    if [ $# -eq 0 ]
+    then
+        _error "_delete_bucket requires at least the bucket name"
+        exit 1
+    fi
+
+    B=$1
+    shift
+    EXTRAS="$*"
+
+    if ! aws s3api head-bucket --bucket $B $EXTRAS 2>&1 | grep -q 404;
+    then
+        _run aws s3 rm --recursive s3://$B $EXTRAS > /dev/null
+        _run aws s3 rb s3://$B $EXTRAS > /dev/null
+    fi  
+}
+
+# Based on https://github.com/alestic/aws-cloudformation-stack-status/blob/parent/aws-cloudformation-stack-status
+# Display cloudformation events and errors
 red_font='\e[0;31m'
 red_background='\e[41m'
 green_font='\e[0;32m'
@@ -89,6 +166,15 @@ no_decoration='\e[0m'
 
 function _describe()
 {
+    if [ $# -ne 2 ]
+    then
+        _error "_describe requires stack name and timestamp"
+        exit 1
+    fi
+
+    STACK=$1
+    TIMESTAMP=$2
+
     rm -f $tmp.failed
     touch $tmp.failed
 
@@ -98,7 +184,7 @@ function _describe()
         --output text \
         --query 'StackEvents[*].[Timestamp,LogicalResourceId,ResourceType,ResourceStatus]' \
     | sed -r -e 's/\.[0-9]{6}\+[0-9]{2}\:[0-9]{2}//g' \
-    | awk -v now="$1" -v failed="$tmp.failed" '{if ($1 > now) { print; if (match($4, "[A-Z_]*FAILED[A-Z_]*")) print >> failed}}' \
+    | awk -v now="$TIMESTAMP" -v failed="$tmp.failed" '{if ($1 > now) { print; if (match($4, "[A-Z_]*FAILED[A-Z_]*")) print >> failed}}' \
     | sort \
     | awk '{printf("%-30s %-30s %-30s\n", $2, $3, $4)}' \
     | perl -ane 'print if !$seen{$F[1]}++' \
@@ -125,23 +211,28 @@ function _describe()
 
 function _status()
 {
+    if [ $# -ne 1 ]
+    then
+        _error "_status requires stack name"
+        exit 1
+    fi
     echo
 
     while true
     do
         STATUS=$(aws cloudformation describe-stacks \
             --region $REGION \
-            --stack-name $STACK \
+            --stack-name $1 \
             --query Stacks[0].StackStatus \
             --output text) 
 
         OLDNOW=$NOW
         _now
-        _describe $OLDNOW
+        _describe $1 $OLDNOW
 
         if [[ $STATUS != *_IN_PROGRESS ]]
         then
-            _describe $NOW            
+            _describe $1 $NOW            
             echo
             break
         fi
@@ -158,68 +249,73 @@ function _now()
 
 function _outputs()
 {
+    if [ $# -ne 1 ]
+    then
+        _error "_outputs requires stack name"
+        exit 1
+    fi
     _run aws cloudformation describe-stacks \
         --region $REGION \
-        --stack-name $STACK \
+        --stack-name $1 \
         --query 'Stacks[*].Outputs' \
         --output table
 }
 
 function _create()
 {
-    _header "Deploying $STACK"
+    if [ $# -le 1 ]
+    then
+        _error "_create requires stack name and template"
+        exit 1
+    fi
+
+    STACK=$1
+    TEMPLATE=$2
+    shift 2
+
+    if [[ ! -f $TEMPLATE ]]
+    then
+        _error "No such template $TEMPLATE"
+    fi
+
+    _header "Creating stack $STACK with $TEMPLATE"
 
     _now
 
     _walk aws cloudformation create-stack \
         --region $REGION \
         --stack-name $STACK \
-        --template-body file://$FIRST \
-        --output text
+        --template-body file://$TEMPLATE \
+        --output text \
+        --capabilities CAPABILITY_AUTO_EXPAND $*
 
-    _status
+    _status $STACK
 
     _walk aws cloudformation wait stack-create-complete \
         --region $REGION \
         --stack-name $STACK
 
-    _outputs
-}
-
-function _update()
-{
-    for t in $ALLBUTFIRST
-    do
-        _header "Updating $STACK with $t"
-
-        _now
-
-        _walk aws cloudformation update-stack \
-            --region $REGION \
-            --stack-name $STACK \
-            --template-body file://$t \
-            --output text \
-            --capabilities CAPABILITY_AUTO_EXPAND
-
-        _status
-
-        _walk aws cloudformation wait stack-update-complete \
-            --region $REGION \
-            --stack-name $STACK
-
-        _outputs
-    done
+    _outputs $STACK
 }
 
 function _apply()
 {
-    TEMPLATE=soln-$1.yaml
+    if [ $# -le 1 ]
+    then
+        _error "_apply requires at least stack name and template"
+        exit 1
+    fi
+
+    STACK=$1
+    TEMPLATE=$2
+    shift 2
+
     if [[ ! -f $TEMPLATE ]]
     then
         _error "No such template $TEMPLATE"
     fi
 
-    _header "Update $STACK with $TEMPLATE"
+    _header "Update stack $STACK with $TEMPLATE"
 
     _now
 
@@ -228,31 +324,63 @@ function _apply()
         --stack-name $STACK \
         --template-body file://$TEMPLATE \
         --output text \
-        --capabilities CAPABILITY_AUTO_EXPAND
+        --capabilities CAPABILITY_AUTO_EXPAND $*
 
-    _status
+    _status $STACK
 
     _walk aws cloudformation wait stack-update-complete \
         --region $REGION \
         --stack-name $STACK
 
-    _outputs
+    _outputs $STACK
+}
+
+
+function _update()
+{
+    if [[ $# -le 1 ]]
+    then
+        _error "_update expecting stack and list of templates"
+    fi
+
+    STACK=$1
+    shift
+
+    _header "Updating $STACK with $*"
+
+    for t in $*
+    do
+        _apply $STACK $t
+    done
 }
 
 function _delete()
 {
-    _header "Deleting $STACK"
+    if [ $# -ne 1 ]
+    then
+        _error "_delete requires stack name"
+        exit 1
+    fi
 
-    _now
+    _header "Deleting stack $1"
 
     _run aws cloudformation delete-stack \
         --region $REGION \
-        --stack-name $STACK
+        --stack-name $1
 
     _run aws cloudformation wait stack-delete-complete \
         --region $REGION \
-        --stack-name $STACK
+        --stack-name $1
 }
+
+function _copy_to_bucket()
+{
+    _header "Create $BUCKET for nested templates"
+
+    _create_bucket $BUCKET --region $REGION
+    _walk aws s3 sync . s3://$BUCKET
+}
+
 
 function _usage()
 {
@@ -267,36 +395,80 @@ then
     _usage
 fi
 
-case $1 in
 
-    create)
-        _create
-        ;;
+case $1 in
     
-    update)
-        _update
+    network)
+        _create $NETWORK $NETWORK_FIRST
+        _update $NETWORK $ALLBUT_NETWORK_FIRST
+        ;;
+
+    secgrp)
+        _create $SECGRP $SECGRP_FIRST
+        _apply $SECGRP secgrp-02.yaml --parameters ParameterKey=SshCidr,ParameterValue=0.0.0.0/0
+        ;;
+        
+    bucket)
+        _copy_to_bucket
+        ;;
+
+    parent|nested)
+        _copy_to_bucket
+        _create $PARENT $PARENT_FIRST --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
+        _apply $PARENT parent-02.yaml --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
         ;;
 
     delete)
-        _delete
+        _delete $PARENT
+        _delete_bucket $BUCKET --region $REGION
+        _delete $SECGRP
+        _delete $NETWORK
         ;;
 
     output)
-        _outputs
-        ;;
-
-    all)
-        _create
-        _update
-        _delete
+        _outputs $NETWORK
+        _outputs $SECGRP
+        _outputs $PARENT
         ;;
 
     status)
-        _status
+        _status $NETWORK
+        _status $SECGRP
+        _status $PARENT
         ;;
 
-    [0-1][0-9])
-        _apply $1
+    deploy)
+        _create $NETWORK $NETWORK_FIRST
+        _update $NETWORK $ALLBUT_NETWORK_FIRST
+        _create $SECGRP $SECGRP_FIRST
+        _apply $SECGRP secgrp-02.yaml --parameters ParameterKey=SshCidr,ParameterValue=0.0.0.0/0
+        _copy_to_bucket
+        _create $PARENT $PARENT_FIRST --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
+        _apply $PARENT parent-02.yaml --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
+        ;;
+    
+    n1)
+        _create $NETWORK $NETWORK_FIRST
+        ;;
+
+    n[2-9])
+        _apply $NETWORK network-0${1:1}.yaml
+        ;;
+
+    s1)
+        _create $SECGRP $SECGRP_FIRST
+        ;;
+
+    s2)
+        _apply $SECGRP secgrp-0${1:1}.yaml --parameters ParameterKey=SshCidr,ParameterValue=0.0.0.0/0
+        ;;
+
+    p1)
+        _create $PARENT $PARENT_FIRST --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
+        ;;
+    
+    p[2-3])
+        _apply $PARENT parent-0${1:1}.yaml --parameters ParameterKey=Bucket,ParameterValue=$BUCKET
         ;;
 
     *)
